requirements.txt
altair==5.0.1
asteval==0.9.31
astropy==5.3.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
bokeh==2.4.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
extra-streamlit-components==0.1.56
fonttools==4.42.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lmfit==1.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpld3==0.5.9
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.24.0
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
pyerfa==2.0.0.3
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scipy==1.11.1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
streamlit-authenticator==0.2.2
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uncertainties==3.1.7
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
xyzservices==2023.7.0
zipp==3.16.2
lime-stable==0.9.85